move projectclass to AixWeather
