+++ AixWeather/project_class.py
