+++ tests/test_DWD_forecast.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from parameterized import parameterized
 
-from project_class import ProjectClassDWDForecast
+from AixWeather.project_class import ProjectClassDWDForecast
 from AixWeather.imports.utils_import import MetaData
 from AixWeather.core_data_format_2_output_file import utils_2output
 from config.definitions import ROOT_DIR
