+++ tests/test_output_functions.py
@@ -12,7 +12,7 @@ import pandas as pd
 from tests import utils_4_tests
 from config.definitions import ROOT_DIR
 from AixWeather.imports.utils_import import MetaData
-from project_class import ProjectClassDWDHistorical
+from AixWeather.project_class import ProjectClassDWDHistorical
 
 
 class BaseOutputFunction(unittest.TestCase):
